requirements.txt
selenium
pandas
webdriver_manager
python-dotenv
